+++ packages/isaacscript-common/src/enums/IsaacScriptCommonFeature2.ts
@@ -2,4 +2,5 @@ export enum IsaacScriptCommonFeature2 {
   CUSTOM_REVIVE = "CustomRevive",
   ESAU_JR_DETECTION = "EsauJrDetection",
   FLIP_DETECTION = "FlipDetection",
+  REORDERED_CALLBACKS = "ReorderedCallbacks",
 }